add exceptions

+++ src/main/java/com/gustcustodio/to_do_list_api/services/ItemService.java
@@ -4,6 +4,10 @@ import com.gustcustodio.to_do_list_api.dtos.ItemDTO;
 import com.gustcustodio.to_do_list_api.entities.Item;
 import com.gustcustodio.to_do_list_api.mappers.ItemMapper;
 import com.gustcustodio.to_do_list_api.repositories.ItemRepository;
+import com.gustcustodio.to_do_list_api.services.exceptions.DatabaseException;
+import com.gustcustodio.to_do_list_api.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -23,7 +27,7 @@ public class ItemService {
 
     @Transactional(readOnly = true)
     public ItemDTO getSingleToDoItem(Long id) {
-        Item item = itemRepository.findById(id).orElseThrow();
+        Item item = itemRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Resource not found"));
         return itemMapper.convertEntityToDto(item);
     }
 
@@ -42,15 +46,26 @@ public class ItemService {
 
     @Transactional
     public ItemDTO updateToDoItem(Long id, ItemDTO itemDTO) {
-        Item item = itemRepository.getReferenceById(id);
-        itemMapper.updateEntityFromDto(itemDTO, item);
-        item = itemRepository.save(item);
-        return itemMapper.convertEntityToDto(item);
+        try {
+            Item item = itemRepository.getReferenceById(id);
+            itemMapper.updateEntityFromDto(itemDTO, item);
+            item = itemRepository.save(item);
+            return itemMapper.convertEntityToDto(item);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
     }
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void deleteToDoItem(Long id) {
-        itemRepository.deleteById(id);
+        if (!itemRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
+        try {
+            itemRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Referential integrity failure");
+        }
     }
 
 }
